requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.30.0
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
lxml==4.9.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.8
opencv-python==4.7.0.72
openpyxl==3.0.10
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pytesseract==0.3.10
pytz==2023.3
redis==4.5.4
requests==2.31.0
screeninfo==0.8.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
yarl==1.9.2
